Switch from page.waitForNavigation({frame}) to frame.waitForNavigation()
