+++ lib/FrameManager.js
@@ -16,6 +16,7 @@
 
 const fs = require('fs');
 const EventEmitter = require('events');
+const {NetworkManager} = require('./NetworkManager');
 const {helper, assert} = require('./helper');
 const {ExecutionContext, JSHandle} = require('./ExecutionContext');
 const ElementHandle = require('./ElementHandle');
@@ -118,7 +119,7 @@ class FrameManager extends EventEmitter {
       return;
     assert(parentFrameId);
     const parentFrame = this._frames.get(parentFrameId);
-    const frame = new Frame(this._client, parentFrame, frameId);
+    const frame = new Frame(this._client, parentFrame, frameId, this);
     this._frames.set(frame._id, frame);
     this.emit(FrameManager.Events.FrameAttached, frame);
   }
@@ -145,7 +146,7 @@ class FrameManager extends EventEmitter {
         frame._id = framePayload.id;
       } else {
         // Initial main frame navigation.
-        frame = new Frame(this._client, null, framePayload.id);
+        frame = new Frame(this._client, null, framePayload.id, this);
       }
       this._frames.set(framePayload.id, frame);
       this._mainFrame = frame;
@@ -256,11 +257,14 @@ class Frame {
    * @param {?Frame} parentFrame
    * @param {string} frameId
    */
-  constructor(client, parentFrame, frameId) {
+  constructor(client, parentFrame, frameId, frameManager) {
     this._client = client;
     this._parentFrame = parentFrame;
     this._url = '';
     this._id = frameId;
+    this._frameManager = frameManager;
+
+    this._networkManager = new NetworkManager(client, frameManager);
 
     /** @type {?Promise<!ElementHandle>} */
     this._documentPromise = null;
@@ -708,6 +712,20 @@ class Frame {
     const timeout = helper.isNumber(options.timeout) ? options.timeout : 30000;
     const polling = options.polling || 'raf';
     return new WaitTask(this, pageFunction, 'function', polling, timeout, ...args).promise;
+  }
+
+  async waitForNavigation(options = {}) {
+    const timeout = typeof options.timeout === 'number' ? options.timeout : 30000;
+    const NavigatorWatcher = require('./NavigatorWatcher');
+    const watcher = new NavigatorWatcher(this._frameManager, this, timeout, options);
+
+    const responses = new Map();
+    const listener = helper.addEventListener(this._networkManager, NetworkManager.Events.Response, response => responses.set(response.url(), response));
+    const error = await watcher.navigationPromise();
+    helper.removeEventListeners([listener]);
+    if (error)
+      throw error;
+    return responses.get(this.url()) || null;
   }
 
   /**
