+++ lib/Page.js
@@ -620,17 +620,8 @@ class Page extends EventEmitter {
    * @return {!Promise<?Puppeteer.Response>}
    */
   async waitForNavigation(options = {}) {
-    const frame = options.frame || this._frameManager.mainFrame();
-    const timeout = typeof options.timeout === 'number' ? options.timeout : this._defaultNavigationTimeout;
-    const watcher = new NavigatorWatcher(this._frameManager, frame, timeout, options);
-
-    const responses = new Map();
-    const listener = helper.addEventListener(this._networkManager, NetworkManager.Events.Response, response => responses.set(response.url(), response));
-    const error = await watcher.navigationPromise();
-    helper.removeEventListeners([listener]);
-    if (error)
-      throw error;
-    return responses.get(frame.url()) || null;
+    options.timeout = typeof options.timeout === 'number' ? options.timeout : this._defaultNavigationTimeout;
+    return await this._frameManager.mainFrame().waitForNavigation(options);
   }
 
   /**
